requirements.txt
# Ollama Pulse Dependencies

# Web scraping & HTTP
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.10

# Machine learning & embeddings
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.3

# Data processing
python-dateutil==2.8.2

# Optional: Rate limiting
ratelimit==2.2.1


# Future-proofing: OLMoTrace for traceability
olmotrace==0.2.1
datasets==2.14.0

# Future-proofing: LangChain for adaptive intelligence
langchain-ollama==0.1.2
langchain-core==0.3.1
langchain-community==0.2.5
pynostr==0.5.0
websocket-client==1.8.0

# Nostr integration (10th data source)
pynostr==0.5.0
websocket-client==1.8.0
